add new file

+++ copy_img.sh
@@ -15,18 +15,23 @@ if [[ $# < 4 ]]; then
     exit
 fi
 
+if [ $tq_server = "tq-data04" ]; then
+    tq_server="tq-data05"
+    echo $tq_server
+fi
+
 tmp1="/home/tq/"$tq_server"/sentinel_GRD/$orbit_number/*"$data_id"*/*_TC_DB.data/*VV*.img"
 tmp2=$dst_dir"/VV/Sigma0_VV_"$data_id".img"
-cp $tmp1 $tmp2
+[ -e "$tmp2" ] || cp $tmp1 $tmp2
 
 tmp1="/home/tq/"$tq_server"/sentinel_GRD/$orbit_number/*"$data_id"*/*_TC_DB.data/*VV*.hdr"
 tmp2=$dst_dir"/VV/Sigma0_VV_"$data_id".hdr"
-cp $tmp1 $tmp2
+[ -e "$tmp2" ] || cp $tmp1 $tmp2
 
 tmp1="/home/tq/"$tq_server"/sentinel_GRD/$orbit_number/*"$data_id"*/*_TC_DB.data/*VH*.img"
 tmp2=$dst_dir"/VH/Sigma0_VH_"$data_id".img"
-cp $tmp1 $tmp2
+[ -e "$tmp2" ] || cp $tmp1 $tmp2
 
 tmp1="/home/tq/"$tq_server"/sentinel_GRD/$orbit_number/*"$data_id"*/*_TC_DB.data/*VH*.hdr"
 tmp2=$dst_dir"/VH/Sigma0_VH_"$data_id".hdr"
-cp $tmp1 $tmp2
+[ -e "$tmp2" ] || cp $tmp1 $tmp2
